+++ blocks/news-feed/news-feed.js
@@ -1,75 +1,5 @@
 import { formatDate, sortDataByDate } from '../../scripts/util.js';
-
-// Pagination related functions
-function createPaginationContainer() {
-  const paginationContainer = document.createElement('div');
-  paginationContainer.className = 'pagination';
-  return paginationContainer;
-}
-
-function updatePagination(totalItems, itemsPerPage, pageNum, onPageChange) {
-  const paginationContainer = document.querySelector('.pagination');
-  if (!paginationContainer) return;
-
-  // Clear existing pagination
-  paginationContainer.innerHTML = '';
-
-  // Calculate total pages
-  const totalPages = Math.ceil(totalItems / itemsPerPage);
-
-  // Determine how many page numbers to show
-  const maxVisiblePages = 8;
-  let startPage;
-  let endPage;
-
-  if (totalPages <= maxVisiblePages) {
-    startPage = 1;
-    endPage = totalPages;
-  } else if (pageNum <= 4) {
-    startPage = 1;
-    endPage = 7;
-  } else if (pageNum >= totalPages - 3) {
-    startPage = totalPages - 6;
-    endPage = totalPages;
-  } else {
-    startPage = pageNum - 3;
-    endPage = pageNum + 3;
-  }
-
-  // Create numbered page links
-  for (let i = startPage; i <= endPage; i += 1) {
-    const pageLink = document.createElement('a');
-    pageLink.href = '#';
-    pageLink.textContent = i.toString();
-    pageLink.className = 'page-link';
-
-    if (i === pageNum) {
-      pageLink.classList.add('current');
-    }
-
-    pageLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      onPageChange(i);
-    });
-
-    paginationContainer.appendChild(pageLink);
-  }
-
-  // Add next page button if not on last page
-  if (pageNum < totalPages) {
-    const nextLink = document.createElement('a');
-    nextLink.href = '#';
-    nextLink.textContent = '>';
-    nextLink.className = 'page-link next';
-
-    nextLink.addEventListener('click', (e) => {
-      e.preventDefault();
-      onPageChange(pageNum + 1);
-    });
-
-    paginationContainer.appendChild(nextLink);
-  }
-}
+import { createPaginationContainer, updatePagination } from '../../scripts/pagination.js';
 
 // Search related functions
 function createSearchHeader(query, count) {
